+++ R/estimate_parameters_bernoulli.R
@@ -6,6 +6,8 @@
 #' @param smooth_marginals Boolean (TRUE/FALSE) which indicates whether Bayesian smoothing of marginal estimates should be used as described in \strong{Pepke, S., Ver Steeg, G. Comprehensive discovery of subsample gene expression components by information explanation: therapeutic implications in cancer. BMC Med Genomics 10, 12 (2017).} \url{https://doi.org/10.1186/s12920-017-0245-6}
 #' @return Returns a 3D array of dimensions (2, n_hidden, dim_hidden) that represent the probability of each dim_visible occurring for each dim_hidden across each n_hidden.
 #'
+#' @keywords internal
+#'
 estimate_parameters_bernoulli <- function(x_i, p_y_given_x_3d, smooth_marginals){
 
     # Get dimensions from p_y_given_x_3d -> in case of missing value may be different to expected